src/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
black==22.1.0
blinker==1.8.2
blue==0.9.1
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==38.0.4
emoji==2.14.0
flake8==4.0.1
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Caching==2.3.0
flask-http-middleware==0.4.2
Flask-HTTPAuth==4.8.0
Flask-Injector==0.15.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-profiler==1.8.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
injector==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.6.1
msgspec==0.18.6
multidict==6.1.0
mypy-extensions==1.0.0
openpyxl==3.1.5
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==2.4.0
PyJWT==2.9.0
python-dotenv==0.21.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
ruff==0.6.7
simplejson==3.19.3
speaklater==1.3
SQLAlchemy==2.0.35
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
WTForms==2.3.3
yarl==1.12.1
